feat: 处理初始化 401 错误
